arm: allow platforms to select ARM_TIMER_SP804
With the introduction of COMPILE_TEST, unloadable drivers will not
be built by default. ARM_TIMER_SP804 falls into this category and
must be selected by the platform.

The a15 and bb BSPs do not need this timer (since they don't select
it), so we drop it from the config fragments to avoid a warning.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>