ocicontainers: cleanup and split fragments
The oci containers fragment was imported from the configuration of
meta-virtualization. But the structure of fragments in meta-virt is
flat, compared to the kernel-cache.

We drop the duplicate lxc fragment, and sync with the main fragment.
We also split out the vswitch and docker fragments so that they can
be more easily found and re-used.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>