tiny: enable CONFIG_FW_LOADER on a per-BSP basis
With updates to the tooling, we now get a warning that base.cfg turns
off CONFIG_FW_LOADER, but that it is enabled for certain -tiny BSPs.

Since we want this warning to indicate if extra functionality has been
enabled, we leave it as-is in base.cfg, but turn the option on per-BSP
to indicate we really do want it on (and silence the warning).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>