tiny: enable CONFIG_FW_LOADER on a per-BSP basis
Source: yocto-kernel-cache
MR: 107831
Type: Integration
Disposition: Merged from yocto-kernel-cache
ChangeID: d9ac6174eaf3117903213ccf7a0937cce4b6b8a4
Description:

With updates to the tooling, we now get a warning that base.cfg turns
off CONFIG_FW_LOADER, but that it is enabled for certain -tiny BSPs.

Since we want this warning to indicate if extra functionality has been
enabled, we leave it as-is in base.cfg, but turn the option on per-BSP
to indicate we really do want it on (and silence the warning).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>