cpumask: Un-inline cpulist_parse for SMP; prepare for ascii helpers
1/4 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: cpumask: Un-inline cpulist_parse for SMP; prepare for ascii helpers
Date: Fri, 18 Dec 2020 13:37:46 -0500

commit 1270d1d31e1674ed7012c3b4090bedf88add898f in paulmck/linux-rcu

In order to support convenience tokens like "all", and "none" and
"last" in CPU lists, we'll have to use string operations and expand
on what is currently a simple wrapper around the underlying bitmap
function call.

Rather than add header dependencies to cpumask.h and code more complex
operations not really appropriate for a header file, we prepare by
simply un-inlining it here and move it to the lib dir alongside the
other more complex cpumask functions.

Since lib/cpumask.c is built conditionally on CONFIG_SMP, and there
are non-SMP callers, we leave the one-line stub behind for that case.
If they want to check "0-0" is a valid range, they can still do it.
In the meantime, we can add the ascii helpers for CONFIG_SMP users.
The use of NR_CPUS vs. CONFIG_SMP is consistent with the existing file.

Aside from an additional exported symbol in the SMP case, no functional
changes are anticipated with this move.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]

2/4 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: cpumask: Make "all" alias global and not just RCU
Date: Fri, 18 Dec 2020 13:37:47 -0500

commit 1270d1d31e1674ed7012c3b4090bedf88add898f in paulmck/linux-rcu

It is probably better that we don't have subsystem specific
abbreviations or aliases for generic CPU list specifications.

Hence we move the "all" from RCU out to lib/ so that it can be
used in any instance where CPU lists are being parsed.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]

3/4 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: cpumask: Add a "none" alias to complement "all"
Date: Fri, 18 Dec 2020 13:37:48 -0500

commit 1270d1d31e1674ed7012c3b4090bedf88add898f in paulmck/linux-rcu

With global support for a CPU list alias of "all", it seems to just make
sense to also trivially extend support for an opposite "none" specifier.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]

4/4 [
Author: Paul Gortmaker
Email: paul.gortmaker@windriver.com
Subject: cpumask: Add "last" alias for cpu list specifications
Date: Fri, 18 Dec 2020 13:37:49 -0500

commit 1270d1d31e1674ed7012c3b4090bedf88add898f in paulmck/linux-rcu

It seems that a common configuration is to use the 1st couple cores
for housekeeping tasks, and or driving a busy peripheral that generates
a lot of interrupts, or something similar.

This tends to leave the remaining ones to form a pool of similarly
configured cores to take on the real workload of interest to the user.

So on machine A - with 32 cores, it could be 0-3 for "system" and then
4-31 being used in boot args like nohz_full=, or rcu_nocbs= as part of
setting up the worker pool of CPUs.

But then newer machine B is added, and it has 48 cores, and so while
the 0-3 part remains unchanged, the pool setup cpu list becomes 4-47.

Deployment would be easier if we could just simply replace 31 and 47
with "last" and let the system substitute in the actual number at boot;
a number that it knows better than we do.

No need to have custom boot args per node, no need to do a trial boot
in order to snoop /proc/cpuinfo and/or /sys/devices/system/cpu - no
more fencepost errors of using 32 and 48 instead of 31 and 47.

A generic token replacement is used to substitute "last" with the
number of CPUs present before handing off to bitmap processing.  But
it could just as easily be used to replace any placeholder token with
any other token or value only known at/after boot.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>