qemumips32: disable CONFIG_COMPACTION
As mentioned in yocto bug: https://bugzilla.yoctoproject.org/show_bug.cgi?id=13992
testimages are failing due to issues with kernel page migration/compaction
function that does not work very well with MIPS kernel HIGHMEM that is needed for 512Mb.

A valid workaround is to disable CONFIG_COMPACTION and the problematic
migration functionality is also disabled.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>