qemuarm: inhibit optimize for size
We explicitly enable optimize_for_performance in this BSP, but the
arch default is optimize_for_size. These are choice options in the
kernel, so they both can't be enabled.

Unless we tell the audit that we don't want optimize_for_size, it
will notice that fact, and warn that it is not set in the final
.config.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>