ktypes/standard: disable obsolete crypto options by default
CONFIG_CRYPTO_USER_API_ENABLE_OBSOLETE is on by default, but really
shouldn't be on in our kernels for a safe set of baseline options.

This is almost never used, so we disable it by default. If something
needs it, they can turn it on in a board specific config.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>