common-pc*/qemux86*: set CONFIG_ATA_PIIX as built-in
Since the IDE options were made obselete in the kernel, and the
PATA driver is the replacement, we haven't had one of the commonly
used qemu boot devices enabled in our kernel by default.

We change CONFIG_ATA_PIIX to built-in, to re-enable use cases that
boot from default qemu 'hardware'.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>