nxp-imx8: Enable CONFIG_IMX_REMOTEPROC config
We have observed that the testcase is hang when doing the CPU
hotplug validation. This issue is disappeared when the m4
powerdomain is active (The side effect of enabling the IMX
remoteproc driver). We are not sure if this is an ATF issue
or hardware issue and since this kernel option is also enabled
in the SDK kernel, so we enable it by default to workaround
this issue.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>