qemuarm64: cleanup for tiny enablement
Move the USB and VIRTIO GPU stuff out of gfx, as those should be handled
in the usb and virtio scc files.  Add the PCI scc for virtio (which
thinks it's PCI on this QEMU setup), and add a tiny.scc/cfg to get tiny
working.

Also, EARLY_PRINTK does not exist for arm64.  So remove those from the
common configs.

Signed-off-by: Jon Mason <jdmason@kudzu.us>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>