treewide: Drop the obsolete GPIO sysfs ABI
The GPIO sysfs ABI has been marked as obsolete 6 years ago [1] and has
been scheduled to be removed in 2020. And then it is restricted to be
only available to expert users by commit [2]. This restriction triggers
config warning on several BSPs due to the CONFIG_EXPERT is not enabled.
Of course we can fix these warning by enabling the CONFIG_EXPERT,
but maybe it is the time to drop this obsolete option.

[1] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=fe95046e960b
[2] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=3b4feb21158f8

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>