vt/conmakehash: improve reproducibility
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@gmail.com
Subject: vt/conmakehash: improve reproducibility
Date: Sun, 10 Jul 2022 21:37:07 -0400

The file generated by conmakehash capture the application
path used to generate the file. While that can be informative,
it varies based on where the kernel was built, as the full
path is captured.

We tweak the application to use a second input as the "capture
name", and then modify the Makefile to pass the basename of
the source, making it reproducible.

This could be improved by using some sort of path mapping,
or the application manipualing argv[1] itself, but for now
this solves the reprodicibility issue.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>