lib/build_OID_registry: fix reproducibility issues
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@gmail.com
Subject: lib/build_OID_registry: fix reproducibility issues
Date: Sun, 10 Jul 2022 22:56:53 -0400

The script build_OID_registry captures the full path of itself
in the generated data. This causes reproduciblity issues as the
path is captured and packaged.

We use the basename of the script instead, and that allows us
to be reprodicible, with slightly less information captured in
the output data (but the generating script can still easily
be found).

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>