tools: use basename to identify file in gen-mach-types
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@gmail.com
Subject: tools: use basename to identify file in gen-mach-types
Date: Wed, 13 Jul 2022 12:18:15 -0400

FILENAME is replaced by the full path to the executing script. If
the script is executed via a fully specified path, that is captured
in the output. Although it doesn't impact the output, it does trigger
reproducibility warnings/errors.

So we introduce a basename() function in the script and use it to
make sure the output file contains only the name of the awk script.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>