pnmtologo: use relocatable file name
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@gmail.com
Subject: pnmtologo: use relocatable file name
Date: Thu, 14 Jul 2022 14:43:46 -0400

The logo generation utility is capturing the source of the logo
in the generated .c file. The source file is absolute (as passed
by make), so the full path is captured.

This makes the source fail reproducibility tests.

We use basename() to just get the source file name, and use
that in the generated .c file.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>