nxp-s32g: add BRIDGE config as built-in
This patch is to fix the following build warning:

  WARNING: linux-yocto-rt-5.15.59+gitAUTOINC+358077fbd3_ee7baf33b6-r0
    do_kernel_configcheck: [kernel config]: specified values did not
    make it into the kernel's final configuration:

    [NOTE]: 'CONFIG_NET_DSA' last val (y) and .config val (m) do not match

Because CONFIG_NET_DSA (S32G sja1105 switch driver needs this config)
depends on CONFIG_BRIDGE, which is a module by default in rt kernel.

Acctually, The CONFIG_BRIDGE has been set to 'y' in bridge.scc. But in
preempt-rt.scc, the bridge.scc is included before the preempt-rt.cfg.
There may be some side effects if we shuffle the include of the bridge.scc
and preempt-rt.cfg. So set CONFIG_BRIDGE to 'y' in nxp-s32g.cfg explicitly,
to fix the issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>