marvell-octeon: Disable the VFIO No-IOMMU support
In the previous version of Marvell SDK, there was a dependency between
the VFIO No-IOMMU support and the ODP. So our customer has explicitly
asked us to enable this option. But this dependency has already been
removed from the latest Marvell SDK, this kernel option is no longer
necessary and may cause security issues, so drop it.

Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>