nxp-s32g: update to compatible with SDK BSP35 rc9
The SDK BSP35 RC9 has been released, and this patch is to keep kernel
configs alighed with it.

In detail, the Dynamic Frequency Scaling is not supported by default,
so drop the ARM_SCMI_CPUFREQ config, to keep align with SDK.

And due to the HSE update, the following configs:
  CONFIG_UIO_NXP_HSE
  CONFIG_UIO_NXP_HSE_MU0
  CONFIG_CRYPTO_DEV_NXP_HSE_MU1
are added to keep the kernel compatible with HSE userspace demos. Or else,
the uio hse device regist will fail, then HSE demos can not work as normal.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>