mconf: fix output of cflags and libraries
1/1 [
Author: Bruce Ashfield
Email: bruce.ashfield@gmail.com
Subject: mconf: fix output of cflags and libraries
Date: Mon, 17 Jul 2023 17:17:55 -0400

commit 3122c84409d578a5df8bcb1 [kconfig: refactor Makefile to reduce
process forks] changes the way that flags are detected. They are
no longer just echo'd and captured, they are written to a file and
later read.

We adjust our CROSS ncurses patch accordingly.

We'll eventually be able to drop this patch, but not quite yet.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
]
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>