docs: add SECURITY.md and README.md
Layers in yocto ecosystem are adding a security file to
document how CVEs and other security issues should be
handled.

The kernel-cache doesn't specifically have executable
code, but it could have configuration options that lead
to security issues, so we document the process if that
happens.

This also renames the README file to a more common
name.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>