x86-64: don't force EDAC support on everyone
Similar to the argument of why we shouldn't force NUMA on everyone, the
9 chip registered ECC RAM type stuff also tends to be found mostly on
larger server type stuff and less so on embedded targets.

We already have a skeleton EDAC feature, so move the features over
there.  One could argue that we might want to separate into arch
specific config fragments, but to me - that seems overkill at this
point in time.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>