powerpc: drop already unlinked crtsaves commits
As can be seen in the scc file, these commits have already been
"unlinked" since v6.3+ kernels, as the issue is resolved upstream.

They obviously will remain on at least several older branches as a
reference, if needed.  We don't need them hanging around any longer on
the master branch.

We leave an empty scc file just in case a future commit is required to
resolve a qemu ppc64 issue, as long as that remains a valid MACHINE.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>