aptiv: update v6.6 kernel cache support
This patch is to add kernel cache support for aptiv cvc bsp of LTS24,
and the related .scc and .cfg files refer to the files of yocto-6.1
branch.

To make it easier for updating to new SDKs, the nxp-s32g.cfg file is
included into cvc's .scc file, and then the different configure settings
can be overridden in cvc's own .cfg file.

Moreover, as Aptiv required, driver options should be "m" as many as possible,
so the related configures are all set to "m".

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>