features/kvm: enable kvm configs for arm64
The CONFIG_VIRTUALIZATION is not enabled for ARM64, but it is
default for X86. And CONFIG_KVM depends on CONFIG_VIRTUALIZATION.

In addition, the CONFIG_KVM is a bool type in ARM64, it could not
be built as a kernel module, so split KVM configs out based the
architecture to standalone files.

Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>