aptiv: enable NVMEM_S32CC_OCOTP
The NVMEM_S32CC_OCOTP configure is needed by TMU sensors, or else, the
TMU driver probe will be added to deffered list and can not work
normally. In result, the sensors command can get nothing messages as
following:
  ~# sensors -f
  No sensors found!

So drop the line:
 "# CONFIG_NVMEM_S32CC_OCOTP is not set"
to enable the configure by default to fix the above issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>