features/kvm: enable kvm configs for arm64
Source: yocto-kernel-cache
MR: 170710
Type: Integration
Disposition: Merged from yocto-kernel-cache
ChangeID: 9a9b7762f82a98952cd54e281b7b140cdad0d264
Description:

The CONFIG_VIRTUALIZATION is not enabled for ARM64, but it is
default for X86. And CONFIG_KVM depends on CONFIG_VIRTUALIZATION.

In addition, the CONFIG_KVM is a bool type in ARM64, it could not
be built as a kernel module, so split KVM configs out based the
architecture to standalone files.

Signed-off-by: Liu Haitao <haitao.liu@windriver.com>
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
Signed-off-by: Bin Lan <bin.lan.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>