cfg/virtio: add DRM frame buffer emulation
Currently, DRM frame buffer is being set by bochs config (for qemu
usage).  If that is removed, we will need this set in the viortio
config.  Alternatively, if another platform wants to use virtio and
doesn't have bochs, then they'll have to set this manually.  Either way,
this is a good change.

Signed-off-by: Jon Mason <jon.mason@arm.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>