iommu: add config dependencies
commit [cfg/intel.scc: Include iommu.scc for Arrowlake boot support]
moves the iommu configuration to a location that is included by more
BSPs and kernel types.

One of those kernel types is the -tiny kernel, which uses a allnoconfig
base. Due to the allnoconfig, there are missing dependencies to allow
the desired options to be enabled in the final .config

Those missing options throw configuration audit warnings

By adding the required dependencies to the fragment, it is now
standalone and can be included from anywhere without configuration
audit warnings being triggered.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>