nxp-imx9: Enable enetc for nxp-imx95
Enable enetc for nxp-imx95. Set TSN and
NET_SCH_TAPRIO to build-in kernel. This
is because imx95 enetc has TSN code by
default. However, since the current TSN
related options are used as modules, if
enetc is also used as a module, this will
affect the NFS network file system mounting.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>