nvidia-orin: adjust ethernet driver config
On NVIDIA Orin platform, it needs to use the NVIDIA out of tree
ethernet driver to support more features. So remove the one
in kernel v6.12. In addition, the UEFI firmware is not able to
support the onboard ethernet device, so enable the USB network
card(For example, Realtek RTL8153 Gigabit Network Adapter) so
that mount net FS directly.

Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>