edac: move apei configs to separate feature
On some ARM platforms that do not support ACPI, including CONFIG_ACPI_APEI
and related Kconfig options triggers kernel_configcheck warnings. To
resolve this, the APEI-related features have been moved to a separate
apei feature. This allows the APEI functionality to be conditionally
included based on whether the target platform supports it, eliminating
the unnecessary warnings on non-ACPI platforms.

Update the Marvell BSPs to include apei.scc in order to retain their
functionality.

Signed-off-by: Jianpeng Chang <jianpeng.chang.cn@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>